+++ components/devfile-page/Projects.module.css
@@ -1,0 +1,40 @@
+.project {
+  border-radius: 6px;
+  padding-left: 0.5rem;
+  padding-top: 0.2rem;
+  padding-bottom: 0.2rem;
+  cursor: default;
+}
+
+.selected {
+  background-color: #000000;
+  color: #ffffff;
+  border-radius: 6px;
+  padding-left: 0.5rem;
+  padding-top: 0.2rem;
+  padding-bottom: 0.2rem;
+  cursor: default;
+}
+
+.selected:hover {
+  background-color: #adabae;
+  color: #ffffff;
+}
+
+.project:hover {
+  background-color: #adabae;
+  color: #ffffff;
+}
+
+.button:active {
+  background-color: black;
+  color: white;
+}
+.select::-webkit-scrollbar {
+  width: 0.5rem;
+}
+
+.select::-webkit-scrollbar-thumb {
+  background-color: darkgrey;
+  border-radius: 0.25rem;
+}
