+++ next-env.d.ts
@@ -1,2 +1,2 @@
-// / <reference types="next" />
-// / <reference types="next/types/global" />
+/// <reference types="next" />
+/// <reference types="next/types/global" />
