+++ package.json
@@ -7,8 +7,7 @@
     "build": "next build && next export",
     "start": "next start",
     "clean": "rm -rf .next/ out/ node_modules/",
-    "cypress:start": "yarn cypress open",
-    "cypress:start-build": "concurrently --names 'CYPRESS,SERVER' --prefix-colors 'yellow,blue' \"yarn cypress:start\" \"yarn start-build\"",
+    "cypress:start": "concurrently --names 'CYPRESS,SERVER' --prefix-colors 'yellow,blue' \"yarn cypress open\" \"yarn build && yarn start\"",
     "typedoc:build": "node_modules/.bin/typedoc --tsconfig .",
     "typedoc:start": "npx serve docs",
     "jest:test": "jest --watchAll --verbose",
